requirements.txt
Django>=4.0,<5.0
pandas>=1.5.0
requests>=2.28.0
beautifulsoup4>=4.11.0
google-generativeai>=0.3.0
gunicorn>=20.1.0
whitenoise>=6.0.0

